Refactoring test classes
